Remove synchronization on CRL in CRLChecker
Motivation:

Versions of BouncyCastle prior to 1.46 had a race in LazyDERSequence. To
avoid that race, CRLChecker synchronized on the CRL. This leads to lock
contention. After upgrading to BouncyCastle 1.46 this workaround is no
longer needed.

Modification:

Remove the synchronization on the CRL.

Result:

Reduced lock contention leads to higher request throughput.